[feat/#1] 로그인 스낵바 수정
